refactor(deepseek-v4): clean up helper and kernel paths
