+++ python/tokenspeed/runtime/configs/deepseek_v4_cache_spec.py
@@ -23,9 +23,127 @@ from tokenspeed.runtime.configs.paged_cache_spec import PagedCacheGroupSpec
 V4_KERNEL_BLOCK_ROWS: int = 64
 V4_SWA_KV_GROUP_ID = "v4.swa_kv"
 V4_INDEXER_COMPRESSOR_STATE_GROUP_ID = "v4.c4a.indexer_compressor_state"
+DEEPSEEK_V4_FP8_MAX = 448.0
+DEEPSEEK_V4_FP8_QUANT_BLOCK = 64
+DEEPSEEK_V4_FP8_INDEXER_BLOCK_SIZE = 128
+DEEPSEEK_V4_FP8_SCALE_BYTES = 4
+DEEPSEEK_V4_MXFP4_BLOCK_SIZE = 32
+DEEPSEEK_V4_MXFP4_SCALE_BYTES = 1
+DEEPSEEK_V4_SPARSE_PREFILL_TOPK_ALIGNMENT = 128
+DEEPSEEK_V4_COMPRESSED_LOGICAL_BLOCK_SIZE = 256
 _COMPRESSOR_STATE_WINDOW_TOKENS = {4: 8, 128: 128}
 _COMPRESSOR_STATE_ROWS_PER_PAGE = {4: 4, 128: 8}
-_COMPRESSED_LOGICAL_BLOCK_SIZE = 256
+
+
+def deepseek_v4_nope_dim(head_dim: int, rope_dim: int) -> int:
+    nope_dim = int(head_dim) - int(rope_dim)
+    if nope_dim <= 0:
+        raise ValueError(f"head_dim={head_dim} must be larger than rope_dim={rope_dim}")
+    return nope_dim
+
+
+def deepseek_v4_swa_token_stride(head_dim: int, rope_dim: int) -> int:
+    return deepseek_v4_nope_dim(head_dim, rope_dim) + int(rope_dim) * 2
+
+
+def deepseek_v4_swa_scale_dim(head_dim: int, rope_dim: int) -> int:
+    nope_dim = deepseek_v4_nope_dim(head_dim, rope_dim)
+    if nope_dim % DEEPSEEK_V4_FP8_QUANT_BLOCK != 0:
+        raise ValueError(
+            "DeepSeek V4 FP8 NoPE dim must be divisible by "
+            f"{DEEPSEEK_V4_FP8_QUANT_BLOCK}, got {nope_dim}"
+        )
+    return nope_dim // DEEPSEEK_V4_FP8_QUANT_BLOCK + 1
+
+
+def deepseek_v4_swa_row_bytes(head_dim: int, rope_dim: int) -> int:
+    return deepseek_v4_swa_token_stride(head_dim, rope_dim) + deepseek_v4_swa_scale_dim(
+        head_dim, rope_dim
+    )
+
+
+def deepseek_v4_indexer_mxfp4_value_bytes(index_head_dim: int) -> int:
+    index_head_dim = int(index_head_dim)
+    if index_head_dim % 2 != 0:
+        raise ValueError(f"MXFP4 index head dim must be even, got {index_head_dim}")
+    return index_head_dim // 2
+
+
+def deepseek_v4_indexer_mxfp4_scale_dim(index_head_dim: int) -> int:
+    index_head_dim = int(index_head_dim)
+    if index_head_dim % DEEPSEEK_V4_MXFP4_BLOCK_SIZE != 0:
+        raise ValueError(
+            "MXFP4 index head dim must be divisible by "
+            f"{DEEPSEEK_V4_MXFP4_BLOCK_SIZE}, got {index_head_dim}"
+        )
+    return (
+        index_head_dim // DEEPSEEK_V4_MXFP4_BLOCK_SIZE * DEEPSEEK_V4_MXFP4_SCALE_BYTES
+    )
+
+
+def deepseek_v4_indexer_mxfp4_row_bytes(index_head_dim: int) -> int:
+    return deepseek_v4_indexer_mxfp4_value_bytes(
+        index_head_dim
+    ) + deepseek_v4_indexer_mxfp4_scale_dim(index_head_dim)
+
+
+def deepseek_v4_indexer_mxfp4_layout_from_row_bytes(
+    row_bytes: int,
+) -> tuple[int, int, int]:
+    row_bytes = int(row_bytes)
+    value_bytes_per_block = DEEPSEEK_V4_MXFP4_BLOCK_SIZE // 2
+    bytes_per_block = value_bytes_per_block + DEEPSEEK_V4_MXFP4_SCALE_BYTES
+    if row_bytes % bytes_per_block != 0:
+        raise ValueError(
+            f"MXFP4 indexer row bytes must be value+scale aligned, got {row_bytes}"
+        )
+    num_blocks = row_bytes // bytes_per_block
+    value_bytes = num_blocks * value_bytes_per_block
+    scale_bytes = num_blocks * DEEPSEEK_V4_MXFP4_SCALE_BYTES
+    index_head_dim = num_blocks * DEEPSEEK_V4_MXFP4_BLOCK_SIZE
+    if deepseek_v4_indexer_mxfp4_scale_dim(index_head_dim) != scale_bytes:
+        raise ValueError(
+            f"invalid MXFP4 indexer row bytes {row_bytes} for "
+            f"index_head_dim={index_head_dim}"
+        )
+    return index_head_dim, value_bytes, scale_bytes
+
+
+def deepseek_v4_indexer_fp8_scale_bytes(index_head_dim: int) -> int:
+    index_head_dim = int(index_head_dim)
+    if index_head_dim % DEEPSEEK_V4_FP8_INDEXER_BLOCK_SIZE != 0:
+        raise ValueError(
+            "FP8 index head dim must be divisible by "
+            f"{DEEPSEEK_V4_FP8_INDEXER_BLOCK_SIZE}, got {index_head_dim}"
+        )
+    return (
+        index_head_dim
+        // DEEPSEEK_V4_FP8_INDEXER_BLOCK_SIZE
+        * DEEPSEEK_V4_FP8_SCALE_BYTES
+    )
+
+
+def deepseek_v4_indexer_fp8_row_bytes(index_head_dim: int) -> int:
+    return int(index_head_dim) + deepseek_v4_indexer_fp8_scale_bytes(index_head_dim)
+
+
+def deepseek_v4_indexer_fp8_layout_from_row_bytes(
+    row_bytes: int,
+) -> tuple[int, int]:
+    row_bytes = int(row_bytes)
+    bytes_per_block = DEEPSEEK_V4_FP8_INDEXER_BLOCK_SIZE + DEEPSEEK_V4_FP8_SCALE_BYTES
+    if row_bytes % bytes_per_block != 0:
+        raise ValueError(
+            f"FP8 indexer row bytes must be value+scale aligned, got {row_bytes}"
+        )
+    index_head_dim = row_bytes // bytes_per_block * DEEPSEEK_V4_FP8_INDEXER_BLOCK_SIZE
+    scale_bytes = deepseek_v4_indexer_fp8_scale_bytes(index_head_dim)
+    if index_head_dim + scale_bytes != row_bytes:
+        raise ValueError(
+            f"invalid FP8 indexer row bytes {row_bytes} for "
+            f"index_head_dim={index_head_dim}"
+        )
+    return index_head_dim, scale_bytes
 
 
 def v4_compressor_state_group_id(ratio: int) -> str:
@@ -51,7 +169,7 @@ def parse_v4_compressor_state_group_id(group_id: str) -> int | None:
 def _compressed_kernel_block_size(ratio: int) -> int:
     if ratio <= 1:
         raise ValueError(f"ratio must be > 1, got {ratio}")
-    return max(1, _COMPRESSED_LOGICAL_BLOCK_SIZE // ratio)
+    return max(1, DEEPSEEK_V4_COMPRESSED_LOGICAL_BLOCK_SIZE // ratio)
 
 
 def _resolve_sliding_window(hf_config: Any) -> int:
@@ -121,10 +239,29 @@ def build_v4_cache_specs(
 
 
 __all__ = [
+    "DEEPSEEK_V4_COMPRESSED_LOGICAL_BLOCK_SIZE",
+    "DEEPSEEK_V4_FP8_MAX",
+    "DEEPSEEK_V4_FP8_INDEXER_BLOCK_SIZE",
+    "DEEPSEEK_V4_FP8_QUANT_BLOCK",
+    "DEEPSEEK_V4_FP8_SCALE_BYTES",
+    "DEEPSEEK_V4_MXFP4_BLOCK_SIZE",
+    "DEEPSEEK_V4_MXFP4_SCALE_BYTES",
+    "DEEPSEEK_V4_SPARSE_PREFILL_TOPK_ALIGNMENT",
     "V4_INDEXER_COMPRESSOR_STATE_GROUP_ID",
     "V4_KERNEL_BLOCK_ROWS",
     "V4_SWA_KV_GROUP_ID",
     "build_v4_cache_specs",
+    "deepseek_v4_indexer_fp8_layout_from_row_bytes",
+    "deepseek_v4_indexer_fp8_row_bytes",
+    "deepseek_v4_indexer_fp8_scale_bytes",
+    "deepseek_v4_indexer_mxfp4_layout_from_row_bytes",
+    "deepseek_v4_indexer_mxfp4_row_bytes",
+    "deepseek_v4_indexer_mxfp4_scale_dim",
+    "deepseek_v4_indexer_mxfp4_value_bytes",
+    "deepseek_v4_nope_dim",
+    "deepseek_v4_swa_row_bytes",
+    "deepseek_v4_swa_scale_dim",
+    "deepseek_v4_swa_token_stride",
     "parse_v4_compressor_state_group_id",
     "v4_compressed_kv_group_id",
     "v4_compressor_state_group_id",
